+++ src/controllers/EmailsController.php
@@ -17,6 +17,7 @@ use craft\commerce\Plugin;
 use craft\commerce\records\Email as EmailRecord;
 use craft\helpers\App;
 use craft\helpers\ArrayHelper;
+use craft\models\Site;
 use yii\base\ErrorException;
 use yii\base\Exception;
 use yii\base\InvalidConfigException;
@@ -103,6 +104,11 @@ class EmailsController extends BaseAdminController
 
         $variables['emailLanguageOptions'] = array_merge($emailLanguageOptions, LocaleHelper::getSiteAndOtherLanguages());
 
+        $variables['emailRenderSiteOptions'] = [
+            null => Craft::t('commerce', 'The site the order was made in.'),
+            ['optgroup' => Craft::t('commerce', 'Sites')],
+        ] + collect(Craft::$app->getSites()->getAllSites())->mapWithKeys(fn(Site $site) => [$site->id => $site->name])->all();
+
         $variables['readOnly'] = $this->isReadOnlyScreen();
 
         return $this->renderTemplate('commerce/settings/emails/_edit', $variables);
@@ -136,6 +142,8 @@ class EmailsController extends BaseAdminController
             $email = new Email();
         }
 
+        $renderSiteId = $this->request->getBodyParam('renderSiteId');
+
         // Shared attributes
         $email->storeId = $storeId;
         $email->name = $this->request->getBodyParam('name');
@@ -151,6 +159,7 @@ class EmailsController extends BaseAdminController
         $pdfId = $this->request->getBodyParam('pdfId');
         $email->pdfId = $pdfId ?: null;
         $email->language = $this->request->getBodyParam('language');
+        $email->renderSiteId = $renderSiteId ? (int)$renderSiteId : null;
         $email->setSenderAddress($this->request->getBodyParam('senderAddress'));
         $email->setSenderName($this->request->getBodyParam('senderName'));