+++ src/elements/Product.php
@@ -466,6 +466,42 @@ class Product extends Element implements HasStoreInterface
     /**
      * @inheritdoc
      */
+    protected function safeActionMenuItems(): array
+    {
+        $actions = parent::safeActionMenuItems();
+
+        if (
+            Craft::$app->getUser()->getIsAdmin() &&
+            Craft::$app->getConfig()->getGeneral()->allowAdminChanges
+        ) {
+            // Product type settings
+            $productTypeEditId = sprintf('edit-product-type-%s', mt_rand());
+            $actions[] = [
+                'id' => $productTypeEditId,
+                'icon' => 'gear',
+                'label' => Craft::t('commerce', 'Product type settings'),
+            ];
+
+            $view = Craft::$app->getView();
+            $view->registerJsWithVars(fn($id, $params) => <<<JS
+(() => {
+  $('#' + $id).on('activate', function() {
+    const params = $params;
+    new Craft.CpScreenSlideout('commerce/product-types/edit-product-type', {params});
+  });
+})();
+JS, [
+                $view->namespaceInputId($productTypeEditId),
+                ['productTypeId' => $this->typeId],
+            ]);
+        }
+
+        return $actions;
+    }
+
+    /**
+     * @inheritdoc
+     */
     protected static function includeSetStatusAction(): bool
     {
         return true;
@@ -860,6 +896,31 @@ class Product extends Element implements HasStoreInterface
     {
         $type = $this->getType();
         return ElementHelper::translationKey($this, $type->productTitleTranslationMethod, $type->productTitleTranslationKeyFormat);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getIsSlugTranslatable(): bool
+    {
+        return ($this->getType()->slugTranslationMethod !== Field::TRANSLATION_METHOD_NONE);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getSlugTranslationDescription(): ?string
+    {
+        return ElementHelper::translationDescription($this->getType()->slugTranslationMethod);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function getSlugTranslationKey(): string
+    {
+        $type = $this->getType();
+        return ElementHelper::translationKey($this, $type->slugTranslationMethod, $type->slugTranslationKeyFormat);
     }
 
     /**
@@ -1388,7 +1449,9 @@ class Product extends Element implements HasStoreInterface
         $view = Craft::$app->getView();
         $productType = $this->getType();
         // Slug
-        $fields[] = $this->slugFieldHtml($static);
+        if ($productType->showSlugField) {
+            $fields[] = $this->slugFieldHtml($static);
+        }
 
         if ($productType->isStructure && $productType->maxLevels !== 1) {
             $fields[] = (function() use ($static, $productType) {
@@ -1938,6 +2001,17 @@ class Product extends Element implements HasStoreInterface
                 ],
             ],
         ];
+    }
+
+    /**
+     * @inheritdoc
+     */
+    protected function previewTargets(): array
+    {
+        return array_map(function($previewTarget) {
+            $previewTarget['label'] = Craft::t('site', $previewTarget['label']);
+            return $previewTarget;
+        }, $this->getType()->previewTargets ?? []);
     }
 
     /**